When dropping a frame, merge its damage region
When queuing a new buffer to a buffer queue that is in async mode (which
an app can set via eglSwapInterval(display, 0)), we drop any buffer
currently in the queue. We'd throw out the dropped buffer's damage
region, which is obviously not ok, and causes broken rendering on
devices that actually use the damage region info. Instead, merge the
dropped buffer's damage region with the new buffer's damage region.

Bug: 136158117
Test: Verified correct behavior using a test app provided by John Reck
in b/136158117.
Test: Verified the Oslo tutorial no longer has broken rendering.

Change-Id: I0b069e650e7bed7c1a36e962e681cc5a86700444