Reland "servicemanager: fix lazy service issues"
This reverts commit 66417657c54cebaeda3be19e539bb17d89172ed4.

This CL is relanding removing a check, that was
actually hit when services die.

Most of the issues here occured in a single thread, and
they were due to inconsistent state in the state machine
around clients, and it appears there were a lot of other
things added which papered over these issues.

I've marked explanations in more detail in the code
itself.

Bug: 264814573
Test: aidl_lazy_test in a loop at 500x speed by
    adjusting timings, running for an hour

Change-Id: Icbd72470dbe91761800761e81893f3ebe343e430