Reduce debuggerd backtrace timeouts in dumpstate
This change reduces the per-process timeouts for dumping backtraces from
5 (Java processes) and 20 (native processes) seconds to 3 seconds. A
large amount of the time spent in dumpstate's DUMP TRACES task can be
attributed to waiting for timeouts to expire, even though
successfully-dumped backtraces typically complete under 1 second.

If the current 5 or 20 second timeout is reached, it's more likely that
the process is blocked on something else, rather than the trace dump
just needing more time to finish.

Test: - Collect a bug report with 'adb shell bugreportz' and verify that
        there aren't any more timed-out trace dumps than usual.
      - Set the timeout to something even larger (60s) and check that
	there are still timed-out trace dumps, and that there are no
	dumps that were successful because of the increased timeout.
Bug: 276451949
Change-Id: I5fe89fdd720af56030ecdf7dd82ccc61a83fe58a
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: DennySPb <dennyspb@gmail.com>